backend/requirements.txt
sqlalchemy
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
passlib[bcrypt]
python-jose
